# Make newDictTable able to be inlined.