The Go inliner doesn't like for loops (currently). This replaces the loop in newDictTable with a fixed number of operations. In the cases where `newDictTable` is being called with a constant `minCapacity`, the compiler is able to do all the constant computation and emit the result directly.

Please make sure that the included comment makes the following bit operations sensible as I can understand if it is a little on the unique side.