@nairb774 2017-01-18
The macro benchmarks are still quite noisy, at least on the machine I have, but it seems to favor a slight improvement, maybe neutral. I'm still trying to get the benchmark tools in a better place. The changed benchmarks on my machine (using some WIP I need to get cleaned up, so the format is a little off from the current benchmarks):
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCallDefaults              246           269           +9.35%
BenchmarkCallKeywords              433           460           +6.24%
BenchmarkCallKwargs                1463          1374          -6.08%
BenchmarkCallNoArgs                251           242           -3.59%
BenchmarkCallPositionalArgs        273           275           +0.73%
BenchmarkCallVarArgs               490           504           +2.86%
BenchmarkDictCompCreate            487612        474021        -2.79%
BenchmarkGeneratorExpCreate        1334          1285          -3.67%
BenchmarkGeneratorExpIterate       124           121           -2.42%
BenchmarkListCompCreate            140092        135565        -3.23%
BenchmarkArithmetic                1039          920           -11.45%
BenchmarkArithmeticParallel10      915           1022          +11.69%
BenchmarkArithmeticParallel11      1048          1033          -1.43%
BenchmarkArithmeticParallel12      1067          1043          -2.25%
BenchmarkArithmeticParallel2       1094          1041          -4.84%
BenchmarkArithmeticParallel3       1066          1053          -1.22%
BenchmarkArithmeticParallel4       1077          1048          -2.69%
BenchmarkArithmeticParallel5       1066          1048          -1.69%
BenchmarkArithmeticParallel6       1072          1044          -2.61%
BenchmarkArithmeticParallel7       1077          1051          -2.41%
BenchmarkArithmeticParallel8       994           1035          +4.12%
BenchmarkArithmeticParallel9       1049          1044          -0.48%
BenchmarkFib                       14746         14555         -1.30%
BenchmarkFibParallel10             17265         17467         +1.17%
BenchmarkFibParallel11             18022         17882         -0.78%
BenchmarkFibParallel12             18486         18273         -1.15%
BenchmarkFibParallel2              15032         15168         +0.90%
BenchmarkFibParallel3              15205         15387         +1.20%
BenchmarkFibParallel4              15542         15186         -2.29%
BenchmarkFibParallel5              15525         15808         +1.82%
BenchmarkFibParallel6              15846         16208         +2.28%
BenchmarkFibParallel7              16301         16361         +0.37%
BenchmarkFibParallel8              16605         16815         +1.26%
BenchmarkFibParallel9              16951         17070         +0.70%
BenchmarkDictGetItem               132           129           -2.27%
BenchmarkDictSetItem               263           260           -1.14%
BenchmarkDictStringOnlyGetItem     136           129           -5.15%
BenchmarkDictStringOnlySetItem     261           249           -4.60%
BenchmarkHashStrCached             115           114           -0.87%
BenchmarkGeneratorCreate           454           445           -1.98%
BenchmarkGeneratorIterate          118           128           +8.47%
BenchmarkListContains10            958           983           +2.61%
BenchmarkListContains3             477           464           -2.73%
BenchmarkListGetItem               110           111           +0.91%
BenchmarkForXRange                 73.1          77.8          +6.43%
BenchmarkWhileCounter              480           490           +2.08%
BenchmarkWhileXRange               411           401           -2.43%
BenchmarkTupleContains10           2374          2403          +1.22%
BenchmarkTupleContains3            1009          1015          +0.59%
BenchmarkTupleGetItem              102           101           -0.98%

benchmark                  old bytes     new bytes     delta
BenchmarkFibParallel10     17331         17330         -0.01%
BenchmarkFibParallel3      17328         17329         +0.01%
```

32 improvements, and 23 non-improvements. Take this with a grain of salt - still quite noisy even over 3 runs a piece.