@trotterdylan 2017-01-18
Thanks for doing this! I agree this is worthwhile.

One suggestion is to mention in the comment that why we are doing the calculation this way, because it's definitely not obvious :\